fix: break detached action outcome cycle
